api/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10,<3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aioredis>=2.0.1",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "authlib>=1.4.0",
    "bcrypt==4.0.1",
    "casbin>=1.37.0",
    "click>=8.1.8",
    "cython>=3.0.11",
    "email-validator>=2.2.0",
    "fastapi-mail>=1.4.2",
    "fastapi-pagination[sqlalchemy]>=0.12.34",
    "fastapi[standard]==0.112.2",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "importlib-metadata>=8.5.0",
    "importlib-resources>=6.5.2",
    "itsdangerous>=2.2.0",
    "limits>=4.0.0",
    "loguru>=0.7.3",
    "lz4>=4.3.3",
    "more-itertools>=10.5.0",
    "orjson>=3.10.13",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.4",
    "pydantic-extra-types>=2.10.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "python-snappy>=0.7.3",
    "pytz>=2024.2",
    "redis[hiredis]>=5.2.1",
    "requests>=2.32.3",
    "sentry-sdk[fastapi]>=2.19.2",
    "setuptools-cythonize>=1.0.7",
    "shortuuid>=1.0.13",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.36",
    "tenacity>=9.0.0",
    "ujson>=5.10.0",
    "uvicorn[standard]>=0.34.0",
    "watchdog>=6.0.0",
    "grpcio==1.69.0",
    "grpcio-tools==1.69.0",
    "python-dateutil>=2.9.0.post0",
    "vnstock>=3.2.6",
    "pycryptodome>=3.23.0",
    "taskiq[orjson]>=0.11.17",
    "taskiq-aio-pika>=0.4.2",
    "taskiq-redis>=1.0.9",
    "taskiq-fastapi>=0.3.5",
]

[tool.uv]
environments = [
    "sys_platform == 'linux'",
]
dev-dependencies = [
    "black",
    "pylint",
    "flake8",
    "isort",
    "mypy",
    "pre-commit",
    "flake8-blind-except",
    "flake8-bandit",
    "flake8-docstrings",
    "flake8-logging-format",
    "flake8-print",
    "autoflake",
    "types-geoip2",
    "types-pytz",
    "types-redis",
    "bandit",
    "loguru-mypy",
    "pytest",
]

[build-system]
requires = ["setuptools-cythonize", "setuptools", "wheel", "pip"]
build-backend = "setuptools.build_meta"
